Add utility `Transformer` to convert object keys (even nested) with support of special cases of keys being the same as a language keyword.
The TypeScript auto-generation will append a `_` if the model has a property matching a language keyword, e.g. `Type` will be internally mapped to `type_`. When building a model from a JSON payload, we need to take that into account, otherwise the property will not be set when doing:

```
new Model({
  type: 'foo'
});
```